+++ +IspEcgFramework/+data/@DataFeatureSet/DataFeatureSet.m
@@ -1,0 +1,15 @@
+classdef DataFeatureSet < handle
+    %DATAFEATURESET Summary of this class goes here
+    %   Detailed explanation goes here
+    
+    properties
+        RecordingFeatures containers.Map = containers.Map();
+    end
+    
+    methods
+        function dataMatrix = generateFeatureMatrix(featureSelector)
+            
+        end
+    end
+    
+end