+++ +IspEcgFramework/+data/@LeadFeatureSet/LeadFeatureSet.m
@@ -1,0 +1,13 @@
+classdef LeadFeatureSet < handle
+    %LEADFEATURESET Summary of this class goes here
+    %   Detailed explanation goes here
+    
+    properties
+        MarkerFeatures containers.Map = containers.Map();
+        LeadFeature
+    end
+    
+    methods
+    end
+    
+end