+++ +IspEcgFramework/+extraction/@DataFeatureSelector/DataFeatureSelector.m
@@ -1,0 +1,13 @@
+classdef (Abstract) DataFeatureSelector
+    %FEATURESELECTOR Summary of this class goes here
+    %   Detailed explanation goes here
+    
+    properties
+    end
+    
+    methods
+        dataMatrix = generateFeatureMatrix(dataFeatureSet);
+    end
+    
+end
+