+++ +IspEcgFramework/@Constants/Constants.m
@@ -1,10 +1,9 @@
 classdef Constants
-    %CONSTANTS Summary of this class goes here
-    %   Detailed explanation goes here
+    % Constants Library wide constants like the version number and date
     
     properties (Constant)
-        Version = string('1.0.0-beta');
-        VersionData = string('2017-03-17');
+        Version = string('2.0.0-beta'); % Version string. This library uses semantic versioning
+        VersionData = string('2017-05-15');
         MakerPointNotSpecified = int32.empty(0,0);
         NotAFeature = -1;
     end